feat: Support Non-hosted skill downloading
