+++ src/askContainer/commands/viewAllSkills.ts
@@ -15,9 +15,15 @@ import hostedSkillMetadata = model.v1.skill.AlexaHosted.HostedSkillMetadata;
 import { Logger } from '../../logger';
 import { loggableAskError } from '../../exceptions';
 
-const HOSTED_SKILL_RETRIEVAL = 'Getting list of Alexa-Hosted skills...';
-const HOSTED_SKILL_NOT_FOUND = 'No Alexa-Hosted skills under profile. Please create one first.';
-const HOSTED_SKILL_RETRIEVAL_FAILED = 'Alexa-Hosted skills retrieval failed';
+
+enum SkillProvisionType {
+    AHS= 'Alexa-hosted',
+    NAH= 'Non-Alexa-hosted'
+}
+const ALL_SKILLS_QUICK_PICK_TITLE = 'All skills';
+const SKILL_RETRIEVAL_PLACEHOLDER = 'Getting list of Alexa skills...';
+const SKILL_NOT_FOUND = 'No Alexa skills under profile. Please create one first.';
+const SKILL_RETRIEVAL_FAILED = 'Alexa skills retrieval failed.';
 
 export class ViewAllSkillsCommand extends AbstractCommand<void> {
     private skillInfoMap: Map<string, SmapiResource<SkillInfo>>;
@@ -39,18 +45,12 @@ export class ViewAllSkillsCommand extends AbstractCommand<void> {
     private async setHostedSkillInfo(): Promise<void> {
         Logger.verbose(`Calling method: ${this.commandName}.setHostedSkillInfo`);
         await this.setSkillsList();
-        
-        // eslint-disable-next-line no-undef
         await Promise.all(this.skillsList.map(async (skill, index) => {
-            let skillMetadata: hostedSkillMetadata | undefined;
             // Adding the timeout to not throttle the API
             await this.timeout(500*index);
-            skillMetadata = await getHostedSkillMetadata(skill.data.skillSummary.skillId ?? '', this.context);
-            if (skillMetadata) {
-                skill.data.isHosted = true;
-            } else {
-                skill.data.isHosted = false;
-            }
+            const skillMetadata: hostedSkillMetadata | undefined = 
+                await getHostedSkillMetadata(skill.data.skillSummary.skillId ?? '', this.context);
+            skill.data.isHosted = skillMetadata !== undefined;
             skill.data.hostedSkillMetadata = skillMetadata;
     }));
     }
@@ -63,13 +63,11 @@ export class ViewAllSkillsCommand extends AbstractCommand<void> {
         if (skillsArray) {
             this.skillsList = [];
             skillsArray = skillsArray?.sort((skill1, skill2) => {
-                if (!skill1.data || !skill2.data) {
+                if (skill1.data === undefined || skill2.data === undefined) {
                     return 0;
                 } else {
-
                     const skill1Date = new Date(skill1.data.lastUpdated!);
                     const skill2Date = new Date(skill2.data.lastUpdated!);
-
                     if (skill1Date > skill2Date) {
                         return -1;
                     } else if(skill1Date < skill2Date) {
@@ -83,20 +81,20 @@ export class ViewAllSkillsCommand extends AbstractCommand<void> {
             skillsArray.forEach(skill => {
                 this.skillsList.push(
                     new SmapiResource<SkillInfo>(
-                        new SkillInfo(skill.data, undefined), 
-                        skill.data.skillId ?? ''));
+                        new SkillInfo(skill.data, undefined), skill.data.skillId ?? ''));
             });
         }
     }
 
     private async setQpItems(
         context: vscode.ExtensionContext, qp: vscode.QuickPick<vscode.QuickPickItem>): Promise<void> {
-        Logger.verbose(`Calling method: ${this.commandName}.setQPItems`);
+        Logger.verbose(`Calling method: ${this.commandName}.setQPItems, args:`, qp);
+        
         qp.items = [];
         qp.busy = true;
         qp.enabled = false;
-        qp.placeholder = HOSTED_SKILL_RETRIEVAL;
-        
+        qp.placeholder = SKILL_RETRIEVAL_PLACEHOLDER;
+        this.skillInfoMap.clear();
         try {
             const cachedSkills = getCachedSkills(context);
             if (!cachedSkills || cachedSkills.length === 0) {
@@ -105,53 +103,43 @@ export class ViewAllSkillsCommand extends AbstractCommand<void> {
             } else {
                 this.skillsList = cachedSkills;
             }
-    
+
             if (this.skillsList.length > 0) {
                 this.skillsList.forEach(skill => {
-                    if (skill.data.isHosted) {
-                        this.skillInfoMap.set(
-                            getSkillNameFromLocales(skill.data.skillSummary.nameByLocale!), 
-                            skill);
-                    }
+                    this.skillInfoMap.set(
+                        getSkillNameFromLocales(skill.data.skillSummary.nameByLocale!), 
+                        skill);
                 });
             }
     
-            const qpItems: Array<vscode.QuickPickItem> = new Array<vscode.QuickPickItem>();
+            const qpItems: vscode.QuickPickItem[] = new Array<vscode.QuickPickItem>();
             this.skillInfoMap.forEach((value: SmapiResource<SkillInfo>, key: string) => {
                 qpItems.push({
                     label: key,
-                    description: value.data.skillSummary.skillId
+                    description: value.data.isHosted === true ? SkillProvisionType.AHS : SkillProvisionType.NAH,
+                    detail: value.data.skillSummary.skillId
                 });
             });
             qp.items = qpItems;
             qp.busy = false;
-            if (qpItems.length > 0) {
-                qp.placeholder = '';
-            } else {
-                qp.placeholder = HOSTED_SKILL_NOT_FOUND;
-            }
+            qp.placeholder = qpItems.length > 0 ? '' : SKILL_NOT_FOUND;
             qp.enabled = true;
         } catch (err) {
             qp.items = [];
             qp.busy = false;
-            qp.placeholder = HOSTED_SKILL_RETRIEVAL_FAILED;
+            qp.placeholder = SKILL_RETRIEVAL_FAILED;
             qp.enabled = true;
-            throw loggableAskError(`${HOSTED_SKILL_RETRIEVAL_FAILED}`, err, true);
+            throw loggableAskError(`${SKILL_RETRIEVAL_FAILED}`, err, true);
         } 
     }
 
-    // eslint-disable-next-line @typescript-eslint/require-await
-    async execute(context: CommandContext): Promise<void> {
-        Logger.debug(`Calling method: ${this.commandName}`);
-        
-        // Since sometimes quickpick is selecting the first item by default
-        // creating quick pick and adding items manually
+    private async skillsQuickPick(context: CommandContext): Promise<void> {
+        Logger.verbose(`Calling method: ${this.commandName}.skillsQuickPick`);
         const allSkillsQP = vscode.window.createQuickPick();
-        allSkillsQP.title = 'Alexa-Hosted skills';
+        allSkillsQP.title = ALL_SKILLS_QUICK_PICK_TITLE;
         allSkillsQP.canSelectMany = false;
         allSkillsQP.matchOnDescription = true;
 
-        // Took from https://github.com/microsoft/vscode/issues/45466#issuecomment-421006659
         allSkillsQP.onDidChangeValue(() => {
             allSkillsQP.activeItems = [];
         });
@@ -159,8 +147,15 @@ export class ViewAllSkillsCommand extends AbstractCommand<void> {
         allSkillsQP.onDidAccept(() => {
             if (allSkillsQP.activeItems.length !== 0) {
                 allSkillsQP.ignoreFocusOut = false;
-                vscode.commands.executeCommand(
-                    'askContainer.skillsConsole.cloneSkill', this.skillInfoMap.get(allSkillsQP.activeItems[0].label));
+                let executeCommand: string;
+                if (allSkillsQP.activeItems[0].description === SkillProvisionType.AHS) {
+                    executeCommand = 'askContainer.skillsConsole.cloneSkill';
+                } else if (allSkillsQP.activeItems[0].description === SkillProvisionType.NAH) {
+                    executeCommand = 'askContainer.skillsConsole.cloneOtherSkill';
+                } else {
+                    return;
+                }
+                void vscode.commands.executeCommand(executeCommand, this.skillInfoMap.get(allSkillsQP.activeItems[0].label));
             }
         });
 
@@ -173,13 +168,17 @@ export class ViewAllSkillsCommand extends AbstractCommand<void> {
         };
         allSkillsQP.buttons = [refreshButton];
         allSkillsQP.onDidTriggerButton(async (e) => {
-            clearCachedSkills(context.extensionContext);
-            this.skillInfoMap.clear();
-            await this.setQpItems(context.extensionContext, allSkillsQP);
+            if (e.tooltip === 'Refresh skills') {
+                clearCachedSkills(context.extensionContext);
+                await this.setQpItems(context.extensionContext, allSkillsQP);
+            }
         });
-
         allSkillsQP.show();
-
         await this.setQpItems(context.extensionContext, allSkillsQP);
+    }
+
+    async execute(context: CommandContext): Promise<void> {
+        Logger.debug(`Calling method: ${this.commandName}`);
+        await this.skillsQuickPick(context);
     }
 }
